requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
openai==0.27.6
Pygments==2.15.1
python-dotenv==1.0.0
requests==2.29.0
rich==13.3.5
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.3.3
yarl==1.9.2
numba
numpy
torch
tqdm
more-itertools
tiktoken==0.3.1
ffmpeg-python==0.2.0
